rename all components to jsx files
