+++ client/package.json
@@ -2,7 +2,7 @@
   "name": "client",
   "version": "1.5.0",
   "private": true,
-  "main": "src/index.js",
+  "main": "src/index.jsx",
   "dependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@reduxjs/toolkit": "^1.9.7",
