+++ client/src/__tests__/App.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import App from "../App.js";
+import App from "../App.jsx";
 
 const root = createRoot(document.getElementById("root"));
 root.render(<App />);
