+++ client/src/_components/icons/CircleHalfIcon.jsx
