+++ client/src/_components/icons/LightIcon.jsx
