+++ client/src/_components/icons/SpinnerIcon.jsx
