+++ client/src/_components/icons/TrashIcon.jsx
