+++ client/src/pages/PrivacyPolicy.jsx
